backend/requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
datetime-truncate==1.1.1
Django==5.2.6
django-admin-charts==1.5.0
django-bower==5.2.0
django-memoize==2.3.1
django-multiselectfield==1.0.1
django-nvd3==0.10.1
djangorestframework==3.16.1
dotenv==0.9.9
fonttools==4.59.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyparsing==3.2.4
python-nvd3==0.16.0
python-slugify==8.0.4
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
urllib3==2.5.0
